Move deprecation.go and ask_for_confirmation.go to command and component pkg respectively
