+++ cli/runtime/command/deprecation.go
@@ -1,7 +1,7 @@
 // Copyright 2021 VMware, Inc. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package cli
+package command
 
 import (
 	"fmt"
