+++ pkg/v1/tkg/cmd/describe_cluster.go
@@ -18,8 +18,8 @@ import (
 	clusterctltree "sigs.k8s.io/cluster-api/cmd/clusterctl/client/tree"
 	crtclient "sigs.k8s.io/controller-runtime/pkg/client"
 
+	"github.com/vmware-tanzu/tanzu-framework/cli/runtime/command"
 	"github.com/vmware-tanzu/tanzu-framework/cli/runtime/component"
-	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/cli"
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/tkg/log"
 	"github.com/vmware-tanzu/tanzu-framework/pkg/v1/tkg/tkgctl"
 )
@@ -97,11 +97,11 @@ func init() {
 	describeClustersCmd.Flags().StringVar(&cd.showOtherConditions, "show-all-conditions", "", " list of comma separated kind or kind/name for which we should show all the object's conditions (all to show conditions for all the objects)")
 
 	describeCmd.Flags().BoolVar(&cd.disableNoEcho, "disable-no-echo", false, "Disable hiding of a MachineInfrastructure and BootstrapConfig when ready condition is true or it has the Status, Severity and Reason of the machine's object")
-	cli.DeprecateFlagWithAlternative(describeCmd, "disable-no-echo", "1.6.0", "--show-details")
+	command.DeprecateFlagWithAlternative(describeCmd, "disable-no-echo", "1.6.0", "--show-details")
 	describeCmd.Flags().BoolVar(&cd.showDetails, "show-details", false, "Show details of MachineInfrastructure and BootstrapConfig when ready condition is true or it has the Status, Severity and Reason of the machine's object")
 
 	describeCmd.Flags().BoolVar(&cd.disableGroupObjects, "disable-grouping", false, "Disable grouping machines when ready condition has the same Status, Severity and Reason")
-	cli.DeprecateFlagWithAlternative(describeCmd, "disable-grouping", "1.6.0", "--show-group-members")
+	command.DeprecateFlagWithAlternative(describeCmd, "disable-grouping", "1.6.0", "--show-group-members")
 	describeCmd.Flags().BoolVar(&cd.showGroupMembers, "show-group-members", false, "Expand machine groups whose ready condition has the same Status, Severity and Reason")
 
 	describeCmd.AddCommand(describeClustersCmd)
